refactor(examples): the typst PDF example takes over the converting-blocks-to-pdf slug
The new exporter's example replaces the react-pdf one at
interoperability/converting-blocks-to-pdf, so existing links show the
current exporter instead of a deprecated one. The react-pdf example moves
to converting-blocks-to-pdf-react-pdf-deprecated (linked from the docs'
deprecated-exporter section). Generated names/titles and the playground
registry regenerated; docs links validated.
