+++ docs/content/docs/features/export/pdf.mdx
@@ -51,10 +51,10 @@ This works out of the box, fully offline: exports match the editor's look,
 and everything needed (the default fonts and the compiler itself) ships
 inside the package. Nothing is fetched from a CDN.
 
-See the [full example](/examples/interoperability/converting-blocks-to-pdf-ua)
+See the [full example](/examples/interoperability/converting-blocks-to-pdf)
 with a live PDF preview below:
 
-<Example name="interoperability/converting-blocks-to-pdf-ua" />
+<Example name="interoperability/converting-blocks-to-pdf" />
 
 ### Customizing the PDF
 
@@ -253,4 +253,5 @@ import {
 
 Note that its mappings are react-pdf mappings; when migrating to the new
 exporter, custom blocks need a [Typst mapping](#custom-mappings--custom-schemas)
-instead.
+instead. The old exporter's example lives on at
+[converting-blocks-to-pdf-react-pdf-deprecated](/examples/interoperability/converting-blocks-to-pdf-react-pdf-deprecated).
