+++ examples/05-interoperability/05-converting-blocks-to-pdf/.bnexample.json
@@ -2,12 +2,14 @@
   "playground": true,
   "docs": true,
   "author": "yousefed",
-  "tags": ["Interoperability"],
+  "tags": ["Interoperability", "Accessibility"],
   "sharedTestDocument": true,
   "dependencies": {
-    "@blocknote/xl-multi-column": "latest",
     "@blocknote/xl-pdf-exporter": "latest",
-    "@react-pdf/renderer": "^4.5.1"
+    "@blocknote/xl-multi-column": "latest",
+    "@blocknote/math-block": "latest",
+    "@blocknote/diagram-block": "latest",
+    "@blocknote/xl-typst-compiler": "latest"
   },
   "pro": true
 }
