+++ examples/05-interoperability/05-converting-blocks-to-pdf/README.md
@@ -1,11 +1,13 @@
-# Exporting documents to PDF (react-pdf, deprecated)
+# Exporting documents to tagged PDF (PDF/UA)
 
-> **Deprecated:** this example uses the react-pdf based exporter
-> (`@blocknote/xl-pdf-exporter/react-pdf`), which is deprecated and will be
-> removed after a few releases. Use the Typst-based `PDFExporter` instead -
-> see the "Exporting documents to tagged PDF (PDF/UA)" example - which
-> produces accessible, tagged PDF/UA-1 output.
+This example exports the current document to an **accessible, tagged PDF/UA-1**
+file using the Typst-powered `@blocknote/xl-pdf-exporter`. Unlike a plain PDF,
+a tagged PDF carries a logical structure tree (headings, paragraphs, lists,
+tables, figures with alt text, links) that screen readers can navigate.
 
-This example exports the current document (all blocks) as a PDF file and downloads it to your computer.
+**Try it out:** Edit the document — the PDF preview updates live. Click
+"Download" to save it, then verify it with a tool like
+[veraPDF](https://verapdf.org/) (`--flavour ua1`) or the Acrobat Tags panel.
 
-**Try it out:** Edit the document and click "Download .pdf" at the top to download the PDF file.
+> The first export downloads the Typst compiler (wasm) and fonts, so it may take
+> a moment. Images render as tagged placeholder figures for now.
