+++ examples/05-interoperability/05-converting-blocks-to-pdf/index.html
@@ -2,7 +2,7 @@
   <head>
     <meta charset="UTF-8" />
     <meta name="viewport" content="width=device-width, initial-scale=1.0" />
-    <title>Exporting documents to PDF (react-pdf, deprecated)</title>
+    <title>Exporting documents to tagged PDF (PDF/UA)</title>
     <script>
       <!-- AUTO-GENERATED FILE, DO NOT EDIT DIRECTLY -->
     </script>
