+++ examples/05-interoperability/05-converting-blocks-to-pdf/package.json
@@ -20,9 +20,11 @@
     "@mantine/hooks": "^9.0.2",
     "react": "^19.2.3",
     "react-dom": "^19.2.3",
-    "@blocknote/xl-multi-column": "latest",
     "@blocknote/xl-pdf-exporter": "latest",
-    "@react-pdf/renderer": "^4.5.1"
+    "@blocknote/xl-multi-column": "latest",
+    "@blocknote/math-block": "latest",
+    "@blocknote/diagram-block": "latest",
+    "@blocknote/xl-typst-compiler": "latest"
   },
   "devDependencies": {
     "@types/react": "^19.2.3",
