+++ examples/05-interoperability/05-converting-blocks-to-pdf/src/styles.css
@@ -32,7 +32,13 @@
   margin-inline: 16px;
 }
 
+/* A real <button> (keyboard-focusable), restyled to match the label text. */
 .view-label-download {
+  background: none;
+  border: none;
+  color: inherit;
+  font: inherit;
+  padding: 0;
   cursor: pointer;
   text-decoration: underline;
 }
