+++ examples/05-interoperability/11-converting-blocks-to-pdf-react-pdf-deprecated/.bnexample.json
@@ -1,0 +1,13 @@
+{
+  "playground": true,
+  "docs": true,
+  "author": "yousefed",
+  "tags": ["Interoperability"],
+  "sharedTestDocument": true,
+  "dependencies": {
+    "@blocknote/xl-multi-column": "latest",
+    "@blocknote/xl-pdf-exporter": "latest",
+    "@react-pdf/renderer": "^4.5.1"
+  },
+  "pro": true
+}
