+++ examples/05-interoperability/11-converting-blocks-to-pdf-react-pdf-deprecated/README.md
@@ -1,0 +1,11 @@
+# Exporting documents to PDF (react-pdf, deprecated)
+
+> **Deprecated:** this example uses the react-pdf based exporter
+> (`@blocknote/xl-pdf-exporter/react-pdf`), which is deprecated and will be
+> removed after a few releases. Use the Typst-based `PDFExporter` instead -
+> see the "Exporting documents to tagged PDF (PDF/UA)" example - which
+> produces accessible, tagged PDF/UA-1 output.
+
+This example exports the current document (all blocks) as a PDF file and downloads it to your computer.
+
+**Try it out:** Edit the document and click "Download .pdf" at the top to download the PDF file.
