+++ examples/05-interoperability/11-converting-blocks-to-pdf-react-pdf-deprecated/package.json
@@ -1,5 +1,5 @@
 {
-  "name": "@blocknote/example-interoperability-converting-blocks-to-pdf-ua",
+  "name": "@blocknote/example-interoperability-converting-blocks-to-pdf-react-pdf-deprecated",
   "description": "AUTO-GENERATED FILE, DO NOT EDIT DIRECTLY",
   "type": "module",
   "private": true,
@@ -20,11 +20,9 @@
     "@mantine/hooks": "^9.0.2",
     "react": "^19.2.3",
     "react-dom": "^19.2.3",
-    "@blocknote/xl-pdf-exporter": "latest",
     "@blocknote/xl-multi-column": "latest",
-    "@blocknote/math-block": "latest",
-    "@blocknote/diagram-block": "latest",
-    "@blocknote/xl-typst-compiler": "latest"
+    "@blocknote/xl-pdf-exporter": "latest",
+    "@react-pdf/renderer": "^4.5.1"
   },
   "devDependencies": {
     "@types/react": "^19.2.3",
