+++ examples/05-interoperability/11-converting-blocks-to-pdf-react-pdf-deprecated/tsconfig.json
