+++ examples/05-interoperability/11-converting-blocks-to-pdf-react-pdf-deprecated/vite-env.d.ts
