--- examples/05-interoperability/11-converting-blocks-to-pdf-ua/.bnexample.json
@@ -1,15 +1,0 @@
-{
-  "playground": true,
-  "docs": true,
-  "author": "yousefed",
-  "tags": ["Interoperability", "Accessibility"],
-  "sharedTestDocument": true,
-  "dependencies": {
-    "@blocknote/xl-pdf-exporter": "latest",
-    "@blocknote/xl-multi-column": "latest",
-    "@blocknote/math-block": "latest",
-    "@blocknote/diagram-block": "latest",
-    "@blocknote/xl-typst-compiler": "latest"
-  },
-  "pro": true
-}
