+++ playground/src/examples.gen.tsx
@@ -1125,22 +1125,24 @@ export const examples = {
           playground: true,
           docs: true,
           author: "yousefed",
-          tags: ["Interoperability"],
+          tags: ["Interoperability", "Accessibility"],
           sharedTestDocument: true,
           dependencies: {
-            "@blocknote/xl-multi-column": "latest",
             "@blocknote/xl-pdf-exporter": "latest",
-            "@react-pdf/renderer": "^4.5.1",
+            "@blocknote/xl-multi-column": "latest",
+            "@blocknote/math-block": "latest",
+            "@blocknote/diagram-block": "latest",
+            "@blocknote/xl-typst-compiler": "latest",
           } as any,
           pro: true,
         },
-        title: "Exporting documents to PDF (react-pdf, deprecated)",
+        title: "Exporting documents to tagged PDF (PDF/UA)",
         group: {
           pathFromRoot: "examples/05-interoperability",
           slug: "interoperability",
         },
         readme:
-          '> **Deprecated:** this example uses the react-pdf based exporter\n> (`@blocknote/xl-pdf-exporter/react-pdf`), which is deprecated and will be\n> removed after a few releases. Use the Typst-based `PDFExporter` instead -\n> see the "Exporting documents to tagged PDF (PDF/UA)" example - which\n> produces accessible, tagged PDF/UA-1 output.\n\nThis example exports the current document (all blocks) as a PDF file and downloads it to your computer.\n\n**Try it out:** Edit the document and click "Download .pdf" at the top to download the PDF file.',
+          'This example exports the current document to an **accessible, tagged PDF/UA-1**\nfile using the Typst-powered `@blocknote/xl-pdf-exporter`. Unlike a plain PDF,\na tagged PDF carries a logical structure tree (headings, paragraphs, lists,\ntables, figures with alt text, links) that screen readers can navigate.\n\n**Try it out:** Edit the document — the PDF preview updates live. Click\n"Download" to save it, then verify it with a tool like\n[veraPDF](https://verapdf.org/) (`--flavour ua1`) or the Acrobat Tags panel.\n\n> The first export downloads the Typst compiler (wasm) and fonts, so it may take\n> a moment. Images render as tagged placeholder figures for now.',
       },
       {
         projectSlug: "converting-blocks-to-docx",
@@ -1264,32 +1266,31 @@ export const examples = {
           "This example exports the current document (all blocks) as HTML and renders it.\n\n**Relevant Docs:**\n\n- [Converting Blocks to HTML](/docs/features/export/html)",
       },
       {
-        projectSlug: "converting-blocks-to-pdf-ua",
-        fullSlug: "interoperability/converting-blocks-to-pdf-ua",
+        projectSlug: "converting-blocks-to-pdf-react-pdf-deprecated",
+        fullSlug:
+          "interoperability/converting-blocks-to-pdf-react-pdf-deprecated",
         pathFromRoot:
-          "examples/05-interoperability/11-converting-blocks-to-pdf-ua",
+          "examples/05-interoperability/11-converting-blocks-to-pdf-react-pdf-deprecated",
         config: {
           playground: true,
           docs: true,
           author: "yousefed",
-          tags: ["Interoperability", "Accessibility"],
+          tags: ["Interoperability"],
           sharedTestDocument: true,
           dependencies: {
-            "@blocknote/xl-pdf-exporter": "latest",
             "@blocknote/xl-multi-column": "latest",
-            "@blocknote/math-block": "latest",
-            "@blocknote/diagram-block": "latest",
-            "@blocknote/xl-typst-compiler": "latest",
+            "@blocknote/xl-pdf-exporter": "latest",
+            "@react-pdf/renderer": "^4.5.1",
           } as any,
           pro: true,
         },
-        title: "Exporting documents to tagged PDF (PDF/UA)",
+        title: "Exporting documents to PDF (react-pdf, deprecated)",
         group: {
           pathFromRoot: "examples/05-interoperability",
           slug: "interoperability",
         },
         readme:
-          'This example exports the current document to an **accessible, tagged PDF/UA-1**\nfile using the Typst-powered `@blocknote/xl-pdf-exporter`. Unlike a plain PDF,\na tagged PDF carries a logical structure tree (headings, paragraphs, lists,\ntables, figures with alt text, links) that screen readers can navigate.\n\n**Try it out:** Edit the document — the PDF preview updates live. Click\n"Download" to save it, then verify it with a tool like\n[veraPDF](https://verapdf.org/) (`--flavour ua1`) or the Acrobat Tags panel.\n\n> The first export downloads the Typst compiler (wasm) and fonts, so it may take\n> a moment. Images render as tagged placeholder figures for now.',
+          '> **Deprecated:** this example uses the react-pdf based exporter\n> (`@blocknote/xl-pdf-exporter/react-pdf`), which is deprecated and will be\n> removed after a few releases. Use the Typst-based `PDFExporter` instead -\n> see the "Exporting documents to tagged PDF (PDF/UA)" example - which\n> produces accessible, tagged PDF/UA-1 output.\n\nThis example exports the current document (all blocks) as a PDF file and downloads it to your computer.\n\n**Try it out:** Edit the document and click "Download .pdf" at the top to download the PDF file.',
       },
     ],
   },
